Enable further modkey customization.
Now the entire `modifier_key` table is available for overwriting or
changing, which lets me add whatever modkeys I want! BWAHAHA

Think about this though: it might be a good idea to have two tables, one
with "custom keys" and one with "defaults," so that the user doesn't
have to redefine the entire modkey table (since Shift and Ctrl are
almost always going to be in the modkey table as S and C, respectively).

yeah, I'll do that next commit.

STATUS ON THAT LAST THING: I don't think it's going to work.  So users
should just do the whole table.